Make k8s access explicitly read only (#1101)
* Repleace client.Client with client.Reader

This is more aligned with the actual use abd prevents accidental mutation of k8s objects.

Signed-off-by: Etai Lev Ran <elevran@gmail.com>

* Replace client.Client with client.Reader

While this is more aligned with the intent, in some cases the hepler
calls unto gateway API code which is expecting a full fledged client.
In those cases, the original parameter type has been retained (instead
of type casting in the call to Gateway API code)

Signed-off-by: Etai Lev Ran <elevran@gmail.com>

---------

Signed-off-by: Etai Lev Ran <elevran@gmail.com>